move sort relevant components to sort dir
