+++ packages/react-bootstrap-table2/src/sort/symbol.js
